+++ provider/plugin_framework.go
@@ -183,6 +183,7 @@ func (p *Provider) resources(_ context.Context) []func() resource.Resource {
 		sourcemap.NewSourceMapResource,
 		importsavedobjects.NewResource,
 		alertingrule.NewResource,
+		dashboard.NewResource,
 		dataview.NewResource,
 		defaultdataview.NewResource,
 		parameter.NewResource,
@@ -245,7 +246,6 @@ func (p *Provider) resources(_ context.Context) []func() resource.Resource {
 
 func (p *Provider) experimentalResources(_ context.Context) []func() resource.Resource {
 	return []func() resource.Resource{
-		dashboard.NewResource,
 		streams.NewResource,
 	}
 }